src/api/routers/chat.py
"""
API endpoints for chat functionality with AI personas.
"""
from fastapi import APIRouter, HTTPException
from typing import Dict, List, Optional
from pydantic import BaseModel

router = APIRouter()

class ChatMessage(BaseModel):
    """Model for chat message exchange."""
    message: str
    history: Optional[List[Dict[str, str]]] = None
    persona_id: Optional[str] = None

@router.post("/send", response_model=Dict[str, str])
async def send_message(chat: ChatMessage):
    """
    Send a message to the AI persona and get a response.
    """
    # TODO: Implement chat logic with persona
    return {
        "response": f"Received your message: {chat.message}",
        "persona_id": chat.persona_id or "default"
    }

@router.get("/history/{conversation_id}", response_model=List[Dict[str, str]])
async def get_chat_history(conversation_id: str):
    """
    Retrieve chat history for a specific conversation.
    """
    # TODO: Implement chat history retrieval
    return [{"role": "system", "content": f"Conversation {conversation_id} history will be shown here"}]
