requirements.txt
gradio
fastapi
uvicorn
python-multipart
fastapi
uvicorn
python-multipart
python-dotenv
PyYAML
pillow
numpy
python-jose[cryptography]
passlib[bcrypt]
